Ignore tests that need AF_UNIX family

+++ graalpython/lib-python/3/test/test_socket.py
@@ -1799,7 +1799,7 @@ class GeneralModuleTests(unittest.TestCase):
             s.bind((support.HOSTv6, 0, 0, 0))
             self._test_socket_fileno(s, socket.AF_INET6, socket.SOCK_STREAM)
 
-        if hasattr(socket, "AF_UNIX"):
+        if hasattr(socket, "AF_UNIX") and sys.implementation.name != 'graalpython':
             tmpdir = tempfile.mkdtemp()
             self.addCleanup(shutil.rmtree, tmpdir)
             s = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
@@ -4083,28 +4083,38 @@ class SendrecvmsgUnixStreamTestBase(SendrecvmsgConnectedBase,
                                     ConnectedStreamTestMixin, UnixStreamBase):
     pass
 
+
+@unittest.skipIf(sys.implementation.name == 'graalpython', 'GR-28433')
 @requireAttrs(socket.socket, "sendmsg")
 @requireAttrs(socket, "AF_UNIX")
 class SendmsgUnixStreamTest(SendmsgStreamTests, SendrecvmsgUnixStreamTestBase):
     pass
 
+
+@unittest.skipIf(sys.implementation.name == 'graalpython', 'GR-28433')
 @requireAttrs(socket.socket, "recvmsg")
 @requireAttrs(socket, "AF_UNIX")
 class RecvmsgUnixStreamTest(RecvmsgTests, RecvmsgGenericStreamTests,
                             SendrecvmsgUnixStreamTestBase):
     pass
 
+
+@unittest.skipIf(sys.implementation.name == 'graalpython', 'GR-28433')
 @requireAttrs(socket.socket, "recvmsg_into")
 @requireAttrs(socket, "AF_UNIX")
 class RecvmsgIntoUnixStreamTest(RecvmsgIntoTests, RecvmsgGenericStreamTests,
                                 SendrecvmsgUnixStreamTestBase):
     pass
 
+
+@unittest.skipIf(sys.implementation.name == 'graalpython', 'GR-28433')
 @requireAttrs(socket.socket, "sendmsg", "recvmsg")
 @requireAttrs(socket, "AF_UNIX", "SOL_SOCKET", "SCM_RIGHTS")
 class RecvmsgSCMRightsStreamTest(SCMRightsTest, SendrecvmsgUnixStreamTestBase):
     pass
 
+
+@unittest.skipIf(sys.implementation.name == 'graalpython', 'GR-28433')
 @requireAttrs(socket.socket, "sendmsg", "recvmsg_into")
 @requireAttrs(socket, "AF_UNIX", "SOL_SOCKET", "SCM_RIGHTS")
 class RecvmsgIntoSCMRightsStreamTest(RecvmsgIntoMixin, SCMRightsTest,
@@ -4278,7 +4288,8 @@ class BasicSocketPairTest(SocketPairTest):
 
     def _check_defaults(self, sock):
         self.assertIsInstance(sock, socket.socket)
-        if hasattr(socket, 'AF_UNIX'):
+        # GR-28433
+        if hasattr(socket, 'AF_UNIX') and sys.implementation.name != 'graalpython':
             self.assertEqual(sock.family, socket.AF_UNIX)
         else:
             self.assertEqual(sock.family, socket.AF_INET)
@@ -5055,6 +5066,7 @@ class TestExceptions(unittest.TestCase):
             sock.setblocking(False)
 
 
+@unittest.skipIf(sys.implementation.name == 'graalpython', 'GR-28433')
 @unittest.skipUnless(sys.platform == 'linux' and hasattr(socket, "AF_UNIX"), 'Linux specific test')
 class TestLinuxAbstractNamespace(unittest.TestCase):
 
@@ -5097,6 +5109,8 @@ class TestLinuxAbstractNamespace(unittest.TestCase):
             s.bind(bytearray(b"\x00python\x00test\x00"))
             self.assertEqual(s.getsockname(), b"\x00python\x00test\x00")
 
+
+@unittest.skipIf(sys.implementation.name == 'graalpython', 'GR-28433')
 @unittest.skipUnless(hasattr(socket, 'AF_UNIX'), 'test needs socket.AF_UNIX')
 class TestUnixDomain(unittest.TestCase):
